cultiveai-backend/app/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
earthengine-api
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
jinja2
email-validator